+++ src/main/java/com/roadtonerdvana/jtelebot/response/json/Update.java
@@ -7,35 +7,38 @@ import org.codehaus.jackson.annotate.JsonProperty;
  * This object represents an incoming update.
  *
  */
-public class Update{
+public class Update {
 	/**
-	 * The update‘s unique identifier. 
-	 * Update identifiers start from a certain positive number and increase sequentially. 
-	 * This ID becomes especially handy if you’re using Webhooks, 
-	 * since it allows you to ignore repeated updates or to restore the correct update sequence,
-	 *  should they get out of order.
+	 * The update‘s unique identifier. Update identifiers start from a certain
+	 * positive number and increase sequentially. This ID becomes especially
+	 * handy if you’re using Webhooks, since it allows you to ignore repeated
+	 * updates or to restore the correct update sequence, should they get out of
+	 * order.
 	 */
 	@JsonProperty("update_id")
-	private Integer updateId;
+	private Long updateId;
 	/**
 	 * Optional. New incoming message of any kind — text, photo, sticker, etc.
 	 */
 	@JsonProperty("message")
 	private Message message;
-	public Integer getUpdateId() {
+
+	public Long getUpdateId() {
 		return updateId;
 	}
-	public void setUpdateId(final Integer updateId) {
+
+	public void setUpdateId(final Long updateId) {
 		this.updateId = updateId;
 	}
+
 	public Message getMessage() {
 		return message;
 	}
+
 	public void setMessage(final Message message) {
 		this.message = message;
 	}
-	
-	
+
 	@Override
 	public int hashCode() {
 		final int prime = 31;
@@ -45,6 +48,7 @@ public class Update{
 				+ ((updateId == null) ? 0 : updateId.hashCode());
 		return result;
 	}
+
 	@Override
 	public boolean equals(Object obj) {
 		if (this == obj)
@@ -66,10 +70,10 @@ public class Update{
 			return false;
 		return true;
 	}
+
 	@Override
 	public String toString() {
 		return "Update [updateId=" + updateId + ", message=" + message + "]";
 	}
-	
 
 }
